♻️ (dashboard): Move formatDate to policyEffectivenessUtils
Relocate formatDate (with its invalid-date guard) out of the panel
component into the utils module, alongside the other policy helpers, so
the panel file exports only its component.

Refs AAASM-4155
