+++ dashboard/src/features/analytics/PolicyEffectivenessPanel.tsx
@@ -7,6 +7,7 @@ import {
   sortRulesByBlocks,
   collectDates,
   ratioToColor,
+  formatDate,
 } from './policyEffectivenessUtils'
 import type { PolicyDay } from './policyEffectivenessUtils'
 
@@ -16,17 +17,6 @@ interface TooltipState {
   day: PolicyDay
   x: number
   y: number
-}
-
-const DATE_FMT = new Intl.DateTimeFormat('en-US', { month: 'short', day: 'numeric' })
-const DATE_PLACEHOLDER = '—'
-
-// A malformed 200 response can carry an unparseable date string; new Date(iso)
-// then yields an Invalid Date and Intl.DateTimeFormat throws "Invalid time
-// value" mid-render. Fall back to a placeholder instead of throwing.
-export function formatDate(iso: string): string {
-  const d = new Date(iso)
-  return Number.isNaN(d.getTime()) ? DATE_PLACEHOLDER : DATE_FMT.format(d)
 }
 
 interface HeatmapCellProps {
